+++ packages/node-core/src/indexer/store.service.spec.ts
@@ -49,4 +49,60 @@ describe('Store Service', () => {
     ];
     expect(() => storeService.isIndexed('MinerIP', 'netUid')).toBeTruthy();
   });
+
+  it('isIndexed support snake case', () => {
+    storeService = new StoreService(null as any, null as any, null as any, null as any);
+    (storeService as any)._modelIndexedFields = [
+      {
+        entityName: 'Transfer',
+        fieldName: 'account_from_id',
+        isUnique: false,
+        type: 'btree',
+      },
+      {
+        entityName: 'Transfer',
+        fieldName: 'account_to_id',
+        isUnique: false,
+        type: 'btree',
+      },
+    ];
+
+    expect(storeService.isIndexed('Transfer', 'account_fromId')).toEqual(true);
+    expect(storeService.isIndexed('Transfer', 'accountToId')).toEqual(true);
+    expect(storeService.isIndexed('Transfer', 'accountToId2')).toEqual(false);
+    expect(storeService.isIndexed('Transfer2', 'accountToId')).toEqual(false);
+  });
+
+  it('isIndexedHistorical support snake case', () => {
+    storeService = new StoreService(null as any, null as any, null as any, null as any);
+    (storeService as any)._historical = false;
+    (storeService as any)._modelIndexedFields = [
+      {
+        entityName: 'Transfer',
+        fieldName: 'account_from_id',
+        isUnique: true,
+        type: 'btree',
+      },
+      {
+        entityName: 'Transfer',
+        fieldName: 'account_to_id',
+        isUnique: true,
+        type: 'btree',
+      },
+      {
+        entityName: 'Transfer',
+        fieldName: 'not_index_id',
+        isUnique: false,
+        type: 'btree',
+      },
+    ];
+
+    expect(storeService.isIndexedHistorical('Transfer', 'account_fromId')).toEqual(true);
+    expect(storeService.isIndexedHistorical('Transfer', 'accountToId')).toEqual(true);
+
+    expect(storeService.isIndexedHistorical('Transfer', 'accountToId2')).toEqual(false);
+    expect(storeService.isIndexedHistorical('Transfer2', 'accountToId')).toEqual(false);
+
+    expect(storeService.isIndexedHistorical('Transfer', 'not_index_id')).toEqual(false);
+  });
 });